Converted to python3
Made several changes to the script that will make it run on python3,
namely all of the calls to `raw_input()` were chagned to `input()` as
python3 removed `raw_input()`. This was done because a lot of the
modules that the script relys on have also moved to python3. Also added
support for copying passwords to clipboard when a user wants to get a
password and also added a prompt which allows the user to choose if they
want colored text in the terminal.